src/third_party/blink/web_tests/platform/mac-retina/external/wpt/event-timing/dblclick-expected.txt

This is a testharness.js-based test.
FAIL Event Timing dblclick. assert_greater_than: The entry's processingStart should be greater than startTime. expected a number greater than 1623.435000000427 but got 1623.435000000427
Harness: the test ran to completion.
